refactor: extract base ingredient picker
